Organize test file and rename.
