+++ map-data/src/test/java/org/triplea/map/data/elements/InfoTest.java
@@ -1,16 +1,15 @@
-package org.triplea.map.reader;
+package org.triplea.map.data.elements;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
-import static org.triplea.map.reader.XmlReaderTestUtils.parseMapXml;
+import static org.triplea.map.data.elements.XmlReaderTestUtils.parseMapXml;
 
 import org.junit.jupiter.api.Test;
-import org.triplea.map.data.elements.Game;
 
-class InfoReadingTest {
+class InfoTest {
   @Test
   void readInfoTag() {
-    final Game parsedMap = parseMapXml("info-tag.xml");
+    final Game parsedMap = parseMapXml("info.xml");
 
     assertThat(parsedMap.getInfo().getName(), is("info-tag-test"));
     assertThat(parsedMap.getInfo().getVersion(), is("123.xyz"));
