+++ map-data/src/test/java/org/triplea/map/data/elements/TripleaTest.java
@@ -1,16 +1,15 @@
-package org.triplea.map.reader;
+package org.triplea.map.data.elements;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
-import static org.triplea.map.reader.XmlReaderTestUtils.parseMapXml;
+import static org.triplea.map.data.elements.XmlReaderTestUtils.parseMapXml;
 
 import org.junit.jupiter.api.Test;
-import org.triplea.map.data.elements.Game;
 
 public class TripleaTest {
   @Test
   void readInfoTag() {
-    final Game parsedMap = parseMapXml("triplea-tag.xml");
+    final Game parsedMap = parseMapXml("triplea.xml");
 
     assertThat(parsedMap.getTriplea().getMinimumVersion(), is("min-version"));
   }
