+++ map-data/src/test/java/org/triplea/map/data/elements/XmlReaderTestUtils.java
@@ -1,4 +1,4 @@
-package org.triplea.map.reader;
+package org.triplea.map.data.elements;
 
 import static com.github.npathai.hamcrestopt.OptionalMatchers.isPresent;
 import static org.hamcrest.MatcherAssert.assertThat;
@@ -8,13 +8,11 @@ import static org.hamcrest.core.Is.is;
 import java.io.IOException;
 import java.io.InputStream;
 import lombok.experimental.UtilityClass;
-import org.triplea.map.data.elements.Game;
+import org.triplea.map.reader.MapElementReader;
+import org.triplea.map.reader.MapReadResult;
 
 @UtilityClass
-public class XmlReaderTestUtils {
-  public static InputStream openFile(final String fileName) {
-    return InfoReadingTest.class.getClassLoader().getResourceAsStream(fileName);
-  }
+class XmlReaderTestUtils {
 
   static Game parseMapXml(final String xmlFileName) {
     try (InputStream stream = openFile(xmlFileName)) {
@@ -26,5 +24,9 @@ public class XmlReaderTestUtils {
     } catch (final IOException e) {
       throw new RuntimeException("Failed to parse (expecting a valid XML file): " + xmlFileName, e);
     }
+  }
+
+  private static InputStream openFile(final String fileName) {
+    return InfoTest.class.getClassLoader().getResourceAsStream(fileName);
   }
 }
