+++ map-data/src/test/resources/game-play.xml
