+++ map-data/src/test/resources/info.xml
