+++ map-data/src/test/resources/relationship-types.xml
