+++ map-data/src/test/resources/resource-list.xml
