+++ map-data/src/test/resources/triplea.xml
